Shuffle things around a bit
