+++ engine/src/history_tables/mod.rs
@@ -10,7 +10,7 @@ use history::{Butterfly, HistoryIndex, HistoryScore};
 use killers::Killers;
 use threats::{ThreatIndex, Threats};
 
-use crate::{evaluate::Score, position::Position, search::params::{cont_corr_weight, material_corr_weight, minor_corr_weight, nonpawn_corr_weight, pawn_corr_weight, MAX_DEPTH}, zobrist::ZHash};
+use crate::{search::params::MAX_DEPTH, zobrist::ZHash};
 
 pub mod corrhist;
 pub mod history;
@@ -172,89 +172,16 @@ impl History {
         self.indices.last().and_then(|&idx| self.countermoves[idx])
     }
 
-    // Killers
-    pub fn add_killer(&mut self, ply: usize, mv: Move) {
-        self.killers[ply].add(mv);
-    }
-
-    pub fn clear_killers(&mut self, ply: usize) {
-        self.killers[ply].clear();
-    }
-
-    pub fn clear_all_killers(&mut self) {
-        self.killers = [Killers::new(); MAX_DEPTH];
-    }
-
     pub fn clear_countermoves(&mut self) {
         self.countermoves = Butterfly::default();
     }
 
-    pub fn clear_nodes(&mut self) {
-        self.node_counts = [[0; Square::COUNT]; Square::COUNT];
-    }
-
+    // Node counter
     pub fn add_nodes(&mut self, mv: Move, nodes: u32) {
         self.node_counts[mv.src()][mv.tgt()] += nodes;
     }
 
     pub fn get_nodes(&self, mv: Move) -> u32 {
         self.node_counts[mv.src()][mv.tgt()]
-    }
-
-    // Corrhist
-    pub fn eval_correction(&self, pos: &Position, ply: usize) -> Score {
-        use Color::*;
-        let us = pos.board.current;
-
-        let pawn_correction = self.corr_hist[us][pos.pawn_hash].corr();
-        let w_nonpawn_correction = self.corr_hist[us][pos.nonpawn_hashes[White]].corr();
-        let b_nonpawn_correction = self.corr_hist[us][pos.nonpawn_hashes[Black]].corr();
-        let material_correction = self.corr_hist[us][pos.material_hash].corr();
-        let minor_correction = self.corr_hist[us][pos.minor_hash].corr();
-        let cont_correction = self
-            .indices
-            .get(ply - 2)
-            .map(|idx| self.contcorr_hist[*idx].corr())
-            .unwrap_or_default();
-
-            let correction =
-              pawn_corr_weight()       * pawn_correction
-            + nonpawn_corr_weight()  * w_nonpawn_correction
-            + nonpawn_corr_weight()  * b_nonpawn_correction
-            + material_corr_weight() * material_correction
-            + minor_corr_weight()    * minor_correction
-            + cont_corr_weight()     * cont_correction;
-
-        correction / 256
-    }
-
-    pub fn update_corrhist(
-        &mut self,
-        pos: &Position,
-        ply: usize,
-        depth: usize,
-        score: Score,
-        eval: Score,
-    ) {
-        use Color::*;
-        let us = pos.board.current;
-
-        // Update the pawn corrhist
-        self.corr_hist[us][pos.pawn_hash].update(score, eval, depth);
-
-        // Update the non-pawn corrhist
-        self.corr_hist[us][pos.nonpawn_hashes[White]].update(score, eval, depth);
-        self.corr_hist[us][pos.nonpawn_hashes[Black]].update(score, eval, depth);
-
-        // Update the material corrhist
-        self.corr_hist[us][pos.material_hash].update(score, eval, depth);
-
-        // Update the minor corrhist
-        self.corr_hist[us][pos.minor_hash].update(score, eval, depth);
-
-        // Update the cont corrhist
-        if let Some(idx) = self.indices.get(ply - 2) {
-            self.contcorr_hist[*idx].update(score, eval, depth);
-        }
     }
 }
